Move codec to types package
