--- x/gov/codec.go
@@ -1,22 +1,0 @@
-package gov
-
-import (
-	"github.com/cosmos/cosmos-sdk/codec"
-)
-
-var msgCdc = codec.New()
-
-// Register concrete types on codec codec
-func RegisterCodec(cdc *codec.Codec) {
-	cdc.RegisterConcrete(MsgSubmitProposal{}, "cosmos-sdk/MsgSubmitProposal", nil)
-	cdc.RegisterConcrete(MsgDeposit{}, "cosmos-sdk/MsgDeposit", nil)
-	cdc.RegisterConcrete(MsgVote{}, "cosmos-sdk/MsgVote", nil)
-
-	cdc.RegisterInterface((*ProposalContent)(nil), nil)
-	cdc.RegisterConcrete(TextProposal{}, "gov/TextProposal", nil)
-	cdc.RegisterConcrete(SoftwareUpgradeProposal{}, "gov/SoftwareUpgradeProposal", nil)
-}
-
-func init() {
-	RegisterCodec(msgCdc)
-}
